+++ changelog.md
@@ -9,6 +9,8 @@
   ([#164](https://github.com/feltcoop/gro/pull/164))
 - make serve task work for production SvelteKit builds
   ([#163](https://github.com/feltcoop/gro/pull/163))
+- add `src/project/gitignore.ts` with `isGitignored` and `loadGitignoreFilter`
+  ([#165](https://github.com/feltcoop/gro/pull/165))
 - add helper `toSvelteKitBasePath` to `src/build/sveltekit.ts`
   ([#163](https://github.com/feltcoop/gro/pull/163))
 - default to some environment variables